Исправление ошибок в организации файлов проекта
