+++ src/main/java/com/simbirsoft/projectManager/rest/dto/response/ProjectResponse.java
@@ -1,4 +1,4 @@
-package com.simbirsoft.projectManager.dto.response;
+package com.simbirsoft.projectManager.rest.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
