+++ src/main/java/com/simbirsoft/projectManager/service/ProjectService.java
@@ -1,7 +1,7 @@
 package com.simbirsoft.projectManager.service;
 
-import com.simbirsoft.projectManager.dto.request.ProjectRequest;
-import com.simbirsoft.projectManager.dto.response.ProjectResponse;
+import com.simbirsoft.projectManager.rest.dto.request.ProjectRequest;
+import com.simbirsoft.projectManager.rest.dto.response.ProjectResponse;
 
 import java.util.UUID;
