+++ src/main/java/com/simbirsoft/projectManager/service/impl/TaskServiceImpl.java
@@ -1,7 +1,7 @@
 package com.simbirsoft.projectManager.service.impl;
 
-import com.simbirsoft.projectManager.dto.request.TaskRequest;
-import com.simbirsoft.projectManager.dto.response.TaskResponse;
+import com.simbirsoft.projectManager.rest.dto.request.TaskRequest;
+import com.simbirsoft.projectManager.rest.dto.response.TaskResponse;
 import com.simbirsoft.projectManager.entity.Task;
 import com.simbirsoft.projectManager.exception.NotFoundException;
 import com.simbirsoft.projectManager.repository.TaskRepository;
