+++ src/main/resources/application.properties
@@ -1,3 +1,6 @@
+#server.servlet.context-path=/
+#server.port=${PORT:8080}
+
 ## default connection pool
 spring.datasource.hikari.connectionTimeout=20000
 spring.datasource.hikari.maximumPoolSize=5
@@ -6,3 +9,6 @@ spring.datasource.hikari.maximumPoolSize=5
 spring.datasource.url=jdbc:postgresql://localhost:5432/management_db
 spring.datasource.username=postgres
 spring.datasource.password=zzzzzzzz
+
+logging.level.root=INFO
+logging.level.org.springframework.security=DEBUG
